🏷️rename: hook => _hook
